Create file for new plugin suppression component
