+++ assets/src/settings-page/plugin-suppression.js
@@ -1,46 +1,52 @@
-
 /**
  * WordPress dependencies
  */
-import { useContext, useEffect, useRef, useState } from '@wordpress/element';
+import { useContext } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import { Options } from '../components/options-context-provider';
-import { Selectable } from '../components/selectable';
 
-/**
- * Plugin suppression section of the settings page.
- */
+export function PluginRow( pluginDetails ) {
+	console.log( pluginDetails );
+
+	return null;
+}
+
 export function PluginSuppression() {
-	const { fetchingOptions } = useContext( Options );
+	const { editedOptions, fetchingOptions } = useContext( Options );
 
-	const [ supportedTemplatesTableFound, setSupportedTemplatesTableFound ] = useState( false );
+	const { suppressible_plugins: suppressiblePlugins } = editedOptions;
 
-	const pluginSuppressionContainer = useRef();
-
-	/**
-	 * Pull the PHP-generated plugin suppression section into the JS-generated area.
-	 */
-	useEffect( () => {
-		const settingsSections = [ ...document.querySelectorAll( '#amp-settings-sections > table' ) ];
-		const supportedTemplatesTable = settingsSections.find( ( section ) => section.querySelector( '.amp-suppressed-plugins' ) );
-		if ( supportedTemplatesTable ) {
-			pluginSuppressionContainer.current.appendChild( supportedTemplatesTable );
-			setSupportedTemplatesTableFound( true );
-		}
-	}, [] );
-
+	console.log( editedOptions );
 	return (
-		<section className={ fetchingOptions || ! supportedTemplatesTableFound ? 'hidden' : '' }>
-			<h2>
-				{ __( 'Plugin Suppression', 'amp' ) }
-			</h2>
-			<Selectable className="plugin-suppression">
-				<div ref={ pluginSuppressionContainer } />
-			</Selectable>
-		</section>
+		<>
+			<p>
+				{ __( 'When a plugin adds markup which is invalid on AMP pages, you have two options: you can review the validation error, determine that the invalid markup is not needed, and let the AMP plugin remove it. Alternatively, you can suppress the offending plugin from running on AMP pages. Below is the list of active plugins which have caused validation issues.', 'amp' ) }
+			</p>
+			<table id="suppressed-plugins-table" className="wp-list-table widefat fixed striped">
+				<thead>
+					<tr>
+						<th className="column-status" scope="col">
+							{ __( 'Status', 'amp' ) }
+						</th>
+						<th className="column-plugin" scope="col">
+							{ __( 'Plugin', 'amp' ) }
+						</th>
+						<th className="column-details" scope="col">
+							{ __( 'Details', 'amp' ) }
+						</th>
+					</tr>
+				</thead>
+				<tbody>
+					{ Object.keys( suppressiblePlugins ).map( ( pluginKey ) => (
+						<PluginRow key={ `plugin-row-${ pluginKey }` } pluginDetails={ suppressiblePlugins[ pluginKey ] } />
+					) ) }
+				</tbody>
+			</table>
+
+		</>
 	);
 }
